fix(acp): Use Go-compatible policy ID generation in HTTP path
The HTTP/CLI path computed policy IDs by hashing raw YAML bytes, while
Go hashes sorted parsed policy fields with a monotonic counter. This
caused Rust and Go to produce different policy IDs for the same policy,
breaking cross-platform ACP tests. Consolidates the Go-compatible
algorithm (already in the FFI path) into acp::policy_yaml and uses it
everywhere. Removes the unused Policy::from_yaml method.
