+++ crates/ffi/src/state/policy_store.rs
@@ -2,7 +2,6 @@ use std::collections::HashMap;
 use std::sync::atomic::{AtomicU64, Ordering};
 
 use parking_lot::RwLock;
-use sha2::{Digest, Sha256};
 
 use crate::policy_yaml::ParsedPolicy;
 
@@ -31,15 +30,9 @@ impl PolicyStore {
     }
 
     /// Add a policy and return its Go-compatible ID.
-    ///
-    /// The ID is generated by:
-    /// 1. Parsing the YAML into a sorted policy structure
-    /// 2. Hashing the parsed fields (name, resources, relations, permissions)
-    /// 3. Combining with a monotonic counter
-    /// 4. SHA256(field_hash || counter_string)
     pub fn add_policy(&self, policy: &str, parsed: &ParsedPolicy) -> String {
         let counter_val = self.counter.fetch_add(1, Ordering::SeqCst);
-        let policy_id = generate_policy_id(parsed, counter_val);
+        let policy_id = acp::policy_yaml::generate_policy_id(parsed, counter_val);
 
         self.policies
             .write()
@@ -64,54 +57,4 @@ impl PolicyStore {
     pub fn list_policies(&self) -> Vec<String> {
         self.policies.read().keys().cloned().collect()
     }
-}
-
-/// Generate a Go-compatible policy ID from parsed policy fields.
-///
-/// Matches Go's `acp_core` `IdTransformer.Transform` + `hashPol`:
-/// 1. Inner hash: SHA256(name + sorted_resources(name + sorted_relations(name) + sorted_permissions(name + expr)))
-/// 2. Outer hash: SHA256(inner_hash_bytes + counter_as_string)
-fn generate_policy_id(policy: &ParsedPolicy, counter: u64) -> String {
-    let inner_hash = hash_policy_fields(policy);
-
-    let mut outer_hasher = Sha256::new();
-    outer_hasher.update(&inner_hash);
-    outer_hasher.update(format!("{}", counter).as_bytes());
-
-    let hash = outer_hasher.finalize();
-    hex::encode(hash)
-}
-
-/// Hash the deterministic parsed fields of a policy.
-///
-/// Matches Go's `hashPol` function: sorts resources, relations, and permissions
-/// by name before hashing.
-fn hash_policy_fields(policy: &ParsedPolicy) -> Vec<u8> {
-    let mut hasher = Sha256::new();
-    hasher.update(policy.name.as_bytes());
-
-    // Sort resources by name (Go's SortTransformer sorts before hashing)
-    let mut resources: Vec<_> = policy.resources.iter().collect();
-    resources.sort_by(|a, b| a.name.cmp(&b.name));
-
-    for resource in resources {
-        hasher.update(resource.name.as_bytes());
-
-        // Sort relations by name
-        let mut relations: Vec<_> = resource.relations.iter().collect();
-        relations.sort_by(|a, b| a.name.cmp(&b.name));
-        for relation in relations {
-            hasher.update(relation.name.as_bytes());
-        }
-
-        // Sort permissions by name
-        let mut permissions: Vec<_> = resource.permissions.iter().collect();
-        permissions.sort_by(|a, b| a.name.cmp(&b.name));
-        for permission in permissions {
-            hasher.update(permission.name.as_bytes());
-            hasher.update(permission.expr.as_bytes());
-        }
-    }
-
-    hasher.finalize().to_vec()
 }
